requirements.txt
<<<<<<< HEAD
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
google-pasta==0.2.0
idna==3.7
imbalanced-learn==0.12.3
iniconfig==2.0.0
joblib==1.4.2
kiwisolver==1.4.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
six==1.16.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
xgboost==2.1.1
=======
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
google-pasta==0.2.0
idna==3.7
imbalanced-learn==0.12.3
iniconfig==2.0.0
joblib==1.4.2
kiwisolver==1.4.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
six==1.16.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
xgboost==2.1.1
>>>>>>> 92e0b791645173a1a776da55050e643a20d94d7d
